Add support for Qt6 and disable Qt5
Required changes to use Qt6 and disable support for Qt5. There are still some adjustments related to Qt6 to be done, but these changes are not required to make Qt6 compile on relevant systems.

* Build system changes Qt6
* Override enterEvent
* Update QKeySequence
* QtChart changes
* Use QScreen to instepct dotsPerInch
* Add app->quit()
* Required updates for code related to QString
* Use RiaQDateTimeTools
* Required changes related to regular expressions
* Support compile on Qt  < 6.5
When version < 6.5 is found, qt_generate_deploy_app_script() is disabled. Compilation of ResInsight will work, but the install target will be incomplete.
* Octave: add missing header.
* Qt Advanced Docking: force Qt6 where both Qt5 and Qt6 is available.

---------

+++ ApplicationLibCode/Application/Tools/RiaColorTools.cpp
@@ -207,15 +207,15 @@ cvf::Color3f RiaColorTools::makeLighter( const cvf::Color3f& color, float normal
 {
     auto qColor = toQColor( color );
 
-    double h = 0.0;
-    double s = 0.0;
-    double l = 0.0;
+    float h = 0.0;
+    float s = 0.0;
+    float l = 0.0;
     qColor.getHslF( &h, &s, &l );
 
     // A negative value will make the color darker
     l = l + ( 1.0 - l ) * normalizedScalingFactor;
 
-    l = std::clamp( l, 0.0, 1.0 );
+    l = std::clamp( l, 0.0f, 1.0f );
 
     qColor.setHslF( h, s, l );
 
@@ -228,10 +228,10 @@ cvf::Color3f RiaColorTools::makeLighter( const cvf::Color3f& color, float normal
 QColor RiaColorTools::modifySaturation( const QColor& color, double factor )
 {
     auto  colorSaturation( color );
-    qreal h, s, v;
+    float h, s, v;
     color.getHsvF( &h, &s, &v );
 
-    s = std::clamp( s * factor, 0.0, 1.0 );
+    s = std::clamp( (float)( s * factor ), 0.0f, 1.0f );
 
     colorSaturation.setHsvF( h, s, v );
     return colorSaturation;
